+++ nf_core/utils.py
@@ -266,6 +266,74 @@ def is_pipeline_directory(wf_path):
             raise UserWarning(warning)
 
 
+# This is the minimal version of Nextflow required to fetch containers with `nextflow inspect`
+NF_INSPECT_MIN_NF_VERSION = (25, 4, 4, False)
+
+# This is the maximal version of nf-core/tools that does not require `nextflow inspect` for downloads
+NFCORE_VER_LAST_WITHOUT_NF_INSPECT = (3, 3, 2)
+
+
+# Pretty print a Nextflow version tuple
+def pretty_nf_version(version: tuple[int, int, int, bool]) -> str:
+    return f"{version[0]}.{version[1]:02}.{version[2]}" + ("-edge" if version[3] else "")
+
+
+def get_nf_version() -> Optional[tuple[int, int, int, bool]]:
+    """Get the version of Nextflow installed on the system."""
+    try:
+        cmd_out = run_cmd("nextflow", "-v")
+        if cmd_out is None:
+            raise RuntimeError("Failed to run Nextflow version check.")
+        out, _ = cmd_out
+        out_str = str(out, encoding="utf-8")  # Ensure we have a string
+
+        version_str = out_str.strip().split()[-1]
+
+        # Check if we are using an edge release
+        is_edge = False
+        edge_split = version_str.split("-")
+        if len(edge_split) > 1:
+            is_edge = True
+            version_str = edge_split[0]
+
+        split_version_str = version_str.split(".")
+        parsed_version_tuple = (
+            int(split_version_str[0]),
+            int(split_version_str[1]),
+            int(split_version_str[2]),
+            is_edge,
+        )
+        return parsed_version_tuple
+    except Exception as e:
+        log.warning(f"Error getting Nextflow version: {e}")
+        return None
+
+
+# Check that the Nextflow version >= the minimal version required
+# This is used to ensure that we can run `nextflow inspect`
+def check_nextflow_version(minimal_nf_version: tuple[int, int, int, bool], silent=False) -> bool:
+    """Check the version of Nextflow installed on the system.
+
+    Args:
+        minimal_nf_version (tuple[int, int, int, bool]): The minimal version of Nextflow required.
+        silent (bool): Whether to log the version or not.
+    Returns:
+        bool: True if the installed version is greater than or equal to `minimal_nf_version`
+    """
+    nf_version = get_nf_version()
+    if nf_version is None:
+        return False
+
+    parsed_version_str = pretty_nf_version(nf_version)
+
+    if silent:
+        log.debug(f"Detected Nextflow version {parsed_version_str}")
+    else:
+        log.info(f"Detected Nextflow version {parsed_version_str}")
+
+    return nf_version >= minimal_nf_version
+
+
 def fetch_wf_config(wf_path: Path, cache_config: bool = True) -> dict:
     """Uses Nextflow to retrieve the the configuration variables
     from a Nextflow workflow.