+++ tests/test_cli.py
@@ -161,6 +161,7 @@ class TestCli(unittest.TestCase):
     @mock.patch("nf_core.pipelines.download.DownloadWorkflow")
     def test_cli_download(self, mock_dl):
         """Test nf-core pipeline is downloaded and cli parameters are passed on."""
+        toplevel_params = {"hide-progress": None}
         params = {
             "revision": "abcdef",
             "outdir": "/path/outdir",
@@ -176,7 +177,12 @@ class TestCli(unittest.TestCase):
             "parallel-downloads": 2,
         }
 
-        cmd = ["pipelines", "download"] + self.assemble_params(params) + ["pipeline_name"]
+        cmd = (
+            self.assemble_params(toplevel_params)
+            + ["pipelines", "download"]
+            + self.assemble_params(params)
+            + ["pipeline_name"]
+        )
         result = self.invoke_cli(cmd)
 
         assert result.exit_code == 0
@@ -195,6 +201,7 @@ class TestCli(unittest.TestCase):
             params["container-cache-utilisation"],
             params["container-cache-index"],
             params["parallel-downloads"],
+            "hide-progress" in toplevel_params,
         )
 
         mock_dl.return_value.download_workflow.assert_called_once()